requirements.txt
pyautogen
chardet
pydantic
pandas
scikit-learn
requests
